+++ mobile/android/android-components/components/browser/session/src/main/java/mozilla/components/browser/session/engine/EngineObserver.kt
@@ -6,6 +6,7 @@ package mozilla.components.browser.session.engine
 
 import android.graphics.Bitmap
 import android.os.Environment
+import androidx.core.net.toUri
 import mozilla.components.browser.session.Download
 import mozilla.components.browser.session.Session
 import mozilla.components.browser.session.engine.request.LoadRequestMetadata
@@ -33,6 +34,9 @@ internal class EngineObserver(
     override fun onLocationChange(url: String) {
         if (session.url != url) {
             session.title = ""
+        }
+
+        if (!isHostEquals(session.url, url)) {
             session.icon = null
         }
 
@@ -47,6 +51,13 @@ internal class EngineObserver(
             it.reject()
             true
         }
+    }
+
+    private fun isHostEquals(sessionUrl: String, newUrl: String): Boolean {
+        val sessionUri = sessionUrl.toUri()
+        val newUri = newUrl.toUri()
+
+        return sessionUri.scheme == newUri.scheme && sessionUri.host == newUri.host
     }
 
     override fun onLoadRequest(url: String, triggeredByRedirect: Boolean, triggeredByWebContent: Boolean) {